test: move some instrumented tests to unit test
